feat!: remove deprecated APIs
